Move entities pkg at server root
