+++ src/server/entities/building.go
