+++ src/server/entities/zombie.go
