+++ src/server/game/surviveler/aidirector.go
@@ -6,9 +6,9 @@ package surviveler
 
 import (
 	"math/rand"
+	"server/entities"
 	"server/events"
 	"server/game"
-	"server/game/entities"
 	"server/math"
 	"time"
